[eSame] bugfix
first bugfix doesn't hit all cases, now it's correct

+++ esame/src/stonefield.py
@@ -82,10 +82,10 @@ class StoneField:
 			self.c[stone.color] += 1
 
 	def Map(self, x, y):
-		return x + y * self.sizex
+		return int(x + y * self.sizex)
 
 	def mark1(self, x, y, force=False):
-		index = int(self.Map(x, y))
+		index = self.Map(x, y)
 
 		if index < 0:
 			self.unmark()
